requirements.txt
Django==3.1.3
django-cors-headers==3.6.0
django-crontab==0.7.1
djangorestframework==3.12.2
gunicorn==20.1.0
progress==1.5
requests==2.22.0
psycopg2==2.8.6
djoser==2.1.0
Pillow==9.0.1
django-simple-history==3.1.1